Change @OneToMany to @ManyToMany to avoid JPA creating unique constraint in usuarios_acesso @JoinTable
-- Why?

It seems that the @OneToMany is understand differently in hibernate context.
Consider you have an Article entity which could have many Tags. Also,
each Article could have many Comments too.

Looking from Comments table perspective. You do not associate one
Comment to more than one Article. Logically, it is not correct. So,
generating the unique constraint seems correct here.

Now, look from tags perspective. It is correct that one Tag could be
assigned to many Articles. So the unique constarint here generate
problem and need to be changed to @ManyToMany

Since it doesn't have any other solution that could be found, the way to
go is to change @OneToMany to @ManyToMany

See this for reference: https://stackoverflow.com/questions/8804572/hibernate-onetomany-and-unique-constraint